% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transcriptomic.R
\docType{data}
\name{Transcriptomic}
\alias{Transcriptomic}
\title{Transcriptomic Expression Matrix for Populus trichocarpa Genotypes}
\format{
A numeric matrix with 199 rows and 25,819 columns:
\describe{
\item{Rows}{Poplar genotypes (e.g., "1-A02", "1-A04", ..., "1-A25")}
\item{Columns}{Gene identifiers from the \emph{Populus trichocarpa} genome v4.1
(e.g., "Potri_001G000700_2_v4_1", "Potri_001G000800_4_v4_1")}
}
}
\source{
...
}
\usage{
Transcriptomic
}
\description{
This dataset contains normalized gene expression levels for 199 genotypes
of \emph{Populus trichocarpa}, grown in a common garden experiment in Orléans, France.
Each value represents the expression level of a specific gene in a given genotype,
based on RNA-seq data.
}
\examples{
data(Transcriptomic)
}
\keyword{datasets}
